requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
supabase>=2.0.0
